taylor/pyproject.toml
[tool.poetry]
name = "research_copilot"
version = "0.1.0"
description = ""
authors = ["Taylor Beck <t@ylor.cc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.3.0"
openai = "^0.28.0"
pandas = "^2.1.0"
numpy = "^1.25.2"
matplotlib = "^3.7.3"
ipykernel = "^6.25.2"
sqlalchemy = "^2.0.20"
tables = "^3.8.0"
psycopg2 = "^2.9.7"
unstructured = {extras = ["all-docs"], version = "^0.10.15"}
python-dotenv = "^1.0.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
slack-bolt = "^1.18.0"
langchain = "^0.0.301"
langchain-experimental = "^0.0.21"
neo4j = "^5.12.0"
wikipedia = "^1.4.0"
sentence-transformers = "^2.2.2"
pdfminer-six = "^20221105"
tiktoken = "^0.5.1"
graphviz = "^0.20.1"
citeurl = "^11.2.1"
spacy = "^3.7.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
